refactor(tests): split provider tests into separate files for maintainability
closes #812
