+++ tests/providers/test_commitizen_provider.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+
+import os
+from pathlib import Path
+from typing import TYPE_CHECKING, Iterator
+
+import pytest
+
+from commitizen.config.base_config import BaseConfig
+from commitizen.providers.commitizen_provider import CommitizenProvider
+
+
+if TYPE_CHECKING:
+    from pytest_mock import MockerFixture
+
+
+@pytest.fixture
+def chdir(tmp_path: Path) -> Iterator[Path]:
+    cwd = Path()
+    os.chdir(tmp_path)
+    yield tmp_path
+    os.chdir(cwd)
+
+
+def test_commitizen_provider(config: BaseConfig, mocker: MockerFixture):
+    config.settings["version"] = "42"
+    mock = mocker.patch.object(config, "set_key")
+
+    provider = CommitizenProvider(config)
+    assert provider.get_version() == "42"
+
+    provider.set_version("43.1")
+    mock.assert_called_once_with("version", "43.1")
